Moved exception classes to their own package.
