+++ src/main/java/io/github/mmbishop/gwttest/core/exceptions/MalformedTestException.java
@@ -14,16 +14,15 @@
  * limitations under the License.
  */
 
-package io.github.mmbishop.gwttest.core;
+package io.github.mmbishop.gwttest.core.exceptions;
 
 /**
- * An exception that is thrown when a GWT test can't be constructed. This typically happens when a test's
- * context class is not public and thus is not accessible to instantiate.
+ * An exception that is thrown when a GWT test is malformed (e.g., given invoked after when).
  */
-public class TestConstructionException extends RuntimeException {
+public class MalformedTestException extends RuntimeException {
 
-    public TestConstructionException(String message, Throwable cause) {
-        super(message, cause);
+    public MalformedTestException(String message) {
+        super(message);
     }
 
 }
