+++ src/test/java/io/github/mmbishop/gwttest/MultipleTestCasesTest.java
@@ -17,7 +17,7 @@
 package io.github.mmbishop.gwttest;
 
 import io.github.mmbishop.gwttest.core.GwtTest;
-import io.github.mmbishop.gwttest.core.MalformedTestException;
+import io.github.mmbishop.gwttest.core.exceptions.MalformedTestException;
 import io.github.mmbishop.gwttest.functions.GwtFunction;
 import io.github.mmbishop.gwttest.functions.GwtFunctionWithArgument;
 import io.github.mmbishop.gwttest.functions.GwtFunctionWithArguments;
