+++ src/test/java/io/github/mmbishop/gwttest/PhaseTransitionTest.java
@@ -17,7 +17,7 @@
 package io.github.mmbishop.gwttest;
 
 import io.github.mmbishop.gwttest.core.GwtTest;
-import io.github.mmbishop.gwttest.core.MalformedTestException;
+import io.github.mmbishop.gwttest.core.exceptions.MalformedTestException;
 import io.github.mmbishop.gwttest.functions.GwtFunction;
 import io.github.mmbishop.gwttest.model.Context;
 import org.junit.jupiter.api.Test;
